Moving common functions to util.h and util.cpp. Adding more comments on util functions.
