+++ demos/TicTacToe/Classes/TicTacToeScene.h
@@ -18,6 +18,8 @@
 #include "MainMenuScene.h"
 #include "TicTacToeLayer.h"
 #include "cocos2d.h"
+#include "util.h"
+
 
 class TicTacToe : public cocos2d::Layer {
  public:
