+++ demos/TicTacToe/Classes/util.cpp
@@ -1,0 +1,65 @@
+// Copyright 2020 Google Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "util.h"
+
+// Logs the message passed in to the console.
+void LogMessage(const char* format, ...) {
+  va_list list;
+  va_start(list, format);
+  vprintf(format, list);
+  va_end(list);
+  printf("\n");
+  fflush(stdout);
+}
+
+// Based on operating system, waits for the specified amount of miliseconds.
+void ProcessEvents(int msec) {
+#ifdef _WIN32
+  Sleep(msec);
+#else
+  usleep(msec * 1000);
+#endif  // _WIN32
+}
+
+// Wait for a Future to be completed. If the Future returns an error, it will
+// be logged.
+void WaitForCompletion(const firebase::FutureBase& future, const char* name) {
+  while (future.status() == firebase::kFutureStatusPending) {
+    ProcessEvents(100);
+  }
+  if (future.status() != firebase::kFutureStatusComplete) {
+    LogMessage("ERROR: %s returned an invalid result.", name);
+  } else if (future.error() != 0) {
+    LogMessage("ERROR: %s returned error %d: %s", name, future.error(),
+               future.error_message());
+  }
+}
+
+// Generates a random uid of a specified length.
+std::string GenerateUid(std::size_t length) {
+  const std::string kCharacters = "0123456789abcdefghjkmnpqrstuvwxyz";
+
+  std::random_device random_device;
+  std::mt19937 generator(random_device());
+  std::uniform_int_distribution<> distribution(0, kCharacters.size() - 1);
+
+  std::string generate_uid;
+
+  for (std::size_t i = 0; i < length; ++i) {
+    generate_uid += kCharacters[distribution(generator)];
+  }
+
+  return generate_uid;
+}
